\name{choresults}
  \docType{data}
  \alias{choresults}
  \title{Consense Result Set for Cho Cell Cycle}
  \description{
    This is a Consense Result set for the Cho Cell Cycle dataset,
    made available to demonstrate some of the post-clustering 
    functionality.
   }
  \usage{data(choresults)}
  \format{A result set, to be utilized by ClusterReport or
    IndexGene.}
  \source{
    \url{http://nhlbi.ohsu.edu/consense/}
  }
  \references{
     Yeung, K.Y., D.R. Haynor, and W.L. Ruzzo, \emph{Validating 
     clustering for gene expression data.} Bioinformatics, 2001. 
     17:309-18.

     Cho, R.J., et al., \emph{A genome-wide transcriptional analysis of 
     the mitotic cell cycle.} Mol Cell, 1998. 2:65-73
  }
  \keyword{datasets}